Update file structure for WebSocket classes
