+++ ChessAI/ChessAI.vcxproj
@@ -136,43 +136,43 @@
     <ClInclude Include="chess.h" />
     <ClInclude Include="chessServer.h" />
     <ClInclude Include="consoleManager.h" />
-    <ClInclude Include="endGameRequest.h" />
-    <ClInclude Include="endGameResponse.h" />
     <ClInclude Include="chessController.h" />
-    <ClInclude Include="getValidMovesRequest.h" />
-    <ClInclude Include="getValidMovesResponse.h" />
-    <ClInclude Include="makeMoveRequest.h" />
-    <ClInclude Include="makeMoveResponse.h" />
-    <ClInclude Include="message.h" />
     <ClInclude Include="move.h" />
-    <ClInclude Include="startGameRequest.h" />
-    <ClInclude Include="startGameResponse.h" />
-    <ClInclude Include="updateClientRequest.h" />
-    <ClInclude Include="updateClientResponse.h" />
     <ClInclude Include="utility.h" />
-    <ClInclude Include="webSocketManager.h" />
+    <ClInclude Include="websocket\dto\endGameRequest.h" />
+    <ClInclude Include="websocket\dto\endGameResponse.h" />
+    <ClInclude Include="websocket\dto\getValidMovesRequest.h" />
+    <ClInclude Include="websocket\dto\getValidMovesResponse.h" />
+    <ClInclude Include="websocket\dto\makeMoveRequest.h" />
+    <ClInclude Include="websocket\dto\makeMoveResponse.h" />
+    <ClInclude Include="websocket\dto\message.h" />
+    <ClInclude Include="websocket\dto\startGameRequest.h" />
+    <ClInclude Include="websocket\dto\startGameResponse.h" />
+    <ClInclude Include="websocket\dto\updateClientRequest.h" />
+    <ClInclude Include="websocket\dto\updateClientResponse.h" />
+    <ClInclude Include="websocket\webSocketManager.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="agent.cpp" />
     <ClCompile Include="BitBoard.cpp" />
     <ClCompile Include="chess.cpp" />
     <ClCompile Include="chessServer.cpp" />
-    <ClCompile Include="endGameRequest.cpp" />
-    <ClCompile Include="endGameResponse.cpp" />
     <ClCompile Include="chessController.cpp" />
-    <ClCompile Include="getValidMovesRequest.cpp" />
-    <ClCompile Include="getValidMovesResponse.cpp" />
     <ClCompile Include="main.cpp" />
-    <ClCompile Include="makeMoveRequest.cpp" />
-    <ClCompile Include="makeMoveResponse.cpp" />
-    <ClCompile Include="message.cpp" />
     <ClCompile Include="move.cpp" />
-    <ClCompile Include="startGameRequest.cpp" />
-    <ClCompile Include="startGameResponse.cpp" />
-    <ClCompile Include="updateClientRequest.cpp" />
-    <ClCompile Include="updateClientResponse.cpp" />
     <ClCompile Include="utility.cpp" />
-    <ClCompile Include="webSocketManager.cpp" />
+    <ClCompile Include="websocket\dto\endGameRequest.cpp" />
+    <ClCompile Include="websocket\dto\endGameResponse.cpp" />
+    <ClCompile Include="websocket\dto\getValidMovesRequest.cpp" />
+    <ClCompile Include="websocket\dto\getValidMovesResponse.cpp" />
+    <ClCompile Include="websocket\dto\makeMoveRequest.cpp" />
+    <ClCompile Include="websocket\dto\makeMoveResponse.cpp" />
+    <ClCompile Include="websocket\dto\message.cpp" />
+    <ClCompile Include="websocket\dto\startGameRequest.cpp" />
+    <ClCompile Include="websocket\dto\startGameResponse.cpp" />
+    <ClCompile Include="websocket\dto\updateClientRequest.cpp" />
+    <ClCompile Include="websocket\dto\updateClientResponse.cpp" />
+    <ClCompile Include="websocket\webSocketManager.cpp" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
