+++ ChessAI/websocket/dto/endGameRequest.cpp
