+++ ChessAI/websocket/dto/endGameRequest.h
