+++ ChessAI/websocket/dto/endGameResponse.cpp
