+++ ChessAI/websocket/dto/endGameResponse.h
