+++ ChessAI/websocket/dto/updateClientRequest.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "message.h"
-#include "chess.h"
+#include "../../utility.h"
 
 namespace boost
 {
@@ -16,23 +16,22 @@ namespace websocket
 	namespace dto
 	{
 		/**
-		 * Message used to request a move.
+		 * Messasge used to update the client.
 		 */
-		struct MakeMoveRequest : Message
+		struct UpdateClientRequest : Message
 		{
-			Color player;
-			PieceNode piece;
-			Position destination;
+			Position source, destination;
 			PieceType promotion;
+			Color nextTurn, winner;
 
-			MakeMoveRequest() = default;
+			UpdateClientRequest() = default;
 
 			/**
-			 * Creates a new instance MakeMoveRequest and populates it using a JSON object.
+			 * Creates a new instance UpdateClientRequest and populates it using a JSON object.
 			 *
 			 * \param json The JSON object used to populate the new instance
 			 */
-			MakeMoveRequest(const boost::json::object& json);
+			UpdateClientRequest(const boost::json::object& json);
 
 			MessageType getMessageType() const override;
